Moved objects files in specials directories
