+++ src/Fabric.cpp
@@ -4,12 +4,12 @@
 #include "Fabric.h"
 #include "GameObjects.h"
 #include "Resources.h"
-#include "objects/Bullet.h"
-#include "objects/Unit.h"
-#include "objects/Shooter.h"
-#include "objects/Ground.h"
-#include "objects/Crystal.h"
-#include "objects/Heart.h"
+#include "objects/interactive/Bullet.h"
+#include "objects/units/Unit.h"
+#include "objects/units/Shooter.h"
+#include "objects/backgrounds/Ground.h"
+#include "objects/interactive/Crystal.h"
+#include "objects/interactive/Heart.h"
 #include "ui/Button.h"
 #include "ui/Text.h"
