+++ src/Scene.cpp
@@ -8,10 +8,10 @@
 #include "GameState.h"
 #include "Application.h"
 #include "Scene.h"
-#include "objects/Ground.h"
-#include "objects/Crystal.h"
-#include "objects/Heart.h"
-#include "objects/Shooter.h"
+#include "objects/backgrounds/Ground.h"
+#include "objects/interactive/Crystal.h"
+#include "objects/interactive/Heart.h"
+#include "objects/units/Shooter.h"
 #include "ui/Button.h"
 #include "Fabric.h"
