+++ src/objects/units/Shooter.cpp
@@ -3,14 +3,14 @@
 #include <cmath>
 #include <SFML/Graphics.hpp>
 #include "Shooter.h"
-#include "Crystal.h"
-#include "Bullet.h"
-#include "Movable.h"
-#include "../CObject.h"
-#include "../Window.h"
-#include "../GameObjects.h"
-#include "../Fabric.h"
-#include "../ui/Text.h"
+#include "../interactive/Crystal.h"
+#include "../interactive/Bullet.h"
+#include "../interfaces/Movable.h"
+#include "../../CObject.h"
+#include "../../Window.h"
+#include "../../GameObjects.h"
+#include "../../Fabric.h"
+#include "../../ui/Text.h"
 
 Shooter::Shooter()
 {
