+++ src/objects/units/Unit.cpp
@@ -3,15 +3,15 @@
 #include <cmath>
 #include <SFML/Graphics.hpp>
 #include "Unit.h"
-#include "Crystal.h"
-#include "Heart.h"
-#include "Bullet.h"
-#include "Movable.h"
-#include "../CObject.h"
-#include "../Window.h"
-#include "../GameObjects.h"
-#include "../Fabric.h"
-#include "../ui/Text.h"
+#include "../interactive/Crystal.h"
+#include "../interactive/Heart.h"
+#include "../interactive/Bullet.h"
+#include "../interfaces/Movable.h"
+#include "../../CObject.h"
+#include "../../Window.h"
+#include "../../GameObjects.h"
+#include "../../Fabric.h"
+#include "../../ui/Text.h"
 
 Unit::Unit()
 {
